--- order-service/src/main/java/io/github/cni274/orderservice/dto/ResultDto.java
@@ -1,9 +1,0 @@
-package io.github.cni274.orderservice.dto;
-
-import lombok.Getter;
-
-@Getter
-public class ResultDto<T> {
-    private T result;
-    private String message;
-}
