--- order-service/src/main/java/io/github/cni274/orderservice/dto/UserDto.java
@@ -1,29 +1,0 @@
-package io.github.cni274.orderservice.dto;
-
-import io.github.cni274.orderservice.entity.User;
-import jakarta.validation.constraints.NotEmpty;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class UserDto {
-    private Long id;
-
-    @NotEmpty
-    private String username;
-    @NotEmpty
-    private String email;
-
-    public static UserDto valueOf(User user) {
-        return UserDto.builder()
-                .id(user.getId())
-                .username(user.getUsername())
-                .email(user.getEmail())
-                .build();
-    }
-}
