--- order-service/src/main/java/io/github/cni274/orderservice/entity/User.java
@@ -1,25 +1,0 @@
-package io.github.cni274.orderservice.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Table(name = "users")
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String username;
-
-    @Column(unique = true, nullable = false)
-    private String email;
-}
