--- order-service/src/main/java/io/github/cni274/orderservice/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package io.github.cni274.orderservice.repository;
-
-import io.github.cni274.orderservice.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-}
