+++ user-service/src/main/java/io/github/cni274/userservice/dto/ResultDto.java
@@ -1,0 +1,9 @@
+package io.github.cni274.userservice.dto;
+
+import lombok.Getter;
+
+@Getter
+public class ResultDto<T> {
+    private T result;
+    private String message;
+}
