+++ user-service/src/main/java/io/github/cni274/userservice/exception/UserException.java
@@ -1,0 +1,11 @@
+package io.github.cni274.userservice.exception;
+
+import io.github.cni274.userservice.enums.ErrorResult;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class UserException extends RuntimeException {
+    private final ErrorResult errorResult;
+}
