+++ user-service/src/main/java/io/github/cni274/userservice/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package io.github.cni274.userservice.repository;
+
+import io.github.cni274.userservice.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+}
