+++ user-service/src/test/java/io/github/cni274/userservice/UserServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package io.github.cni274.userservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
